Break out `GadgetChain` from the `GadgetChainDiscovery` class
This will allow for us to do other things with the discovered chains
rather than just printing them out.
